# Plugins/BTCPayServer.Plugins.Tando/Resources/tando-openapi.yaml
openapi: 3.0.3
info:
  title: Tando BTCPay Plugin API
  description: >
    Merchant onboarding, subscriptions, products, employees, and BTC/M-Pesa split
    payments for the Tando BTCPay Server plugin. All routes are Greenfield-style
    endpoints under `/plugins/api/tando`, authenticated with a BTCPay API key.
  version: "0.1.0"
servers:
  - url: https://{host}/plugins/api/tando
    variables:
      host:
        default: your-btcpay-instance.example.com
security:
  - ApiKey: []
...
paths:
  /stores/{storeId}/mpesa/pay:
    post:
      summary: Record and initiate a KES collection using a stable store/order reference
      description: Duplicate identical requests reuse the payment. A successful HTTP response is a workflow record, not proof of settlement.
      security: [{ ApiKey: [CanModifyStoreSettings] }]
      parameters:
        - $ref: "#/components/parameters/StoreId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [customerPhone, amountKes, orderId]
              properties:
                customerPhone: { type: string }
                amountKes: { type: integer, minimum: 1 }
                orderId: { type: string, maxLength: 200 }
      responses:
        "200": { description: Current persisted collection and disbursement states }
        "400": { description: Invalid request }
        "401": { description: Authentication required }
        "403": { description: Store permission required }
        "404": { description: Store not found }
        "409": { description: Order reused with different payment details }
        "503": { description: Merchant destination missing or provider contract not ready }
  /stores/{storeId}/mpesa/settings:
    put:
      summary: Configure the merchant KES destination
      security: [{ ApiKey: [CanModifyStoreSettings] }]
      parameters:
        - $ref: "#/components/parameters/StoreId"
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/TandoMpesaSettingsRequest" }
      responses:
        "200": { description: Destination saved }
        "400": { description: Destination required }
        "404": { description: Store not found }
  /stores/{storeId}/mpesa/status:
    get:
      summary: Read active KES provider readiness
      security: [{ ApiKey: [CanModifyStoreSettings] }]
      parameters:
        - $ref: "#/components/parameters/StoreId"
      responses:
        "200": { description: Current state or simulated delivery result; HTTP success alone does not mean payment success }
        "401": { description: Authentication required }
        "403": { description: Store permission required }
        "404": { description: Payment not found or mock mode disabled }
        "409": { description: Conflicting identity, amount, destination, or terminal state }
  /stores/{storeId}/mpesa/payments/{orderId}:
    get:
      summary: Read the durable KES payment record
      security: [{ ApiKey: [CanModifyStoreSettings] }]
      parameters:
        - $ref: "#/components/parameters/StoreId"
        - name: orderId
          in: path
          required: true
          schema: { type: string }
      responses:
        "200": { description: Current state or simulated delivery result; HTTP success alone does not mean payment success }
        "401": { description: Authentication required }
        "403": { description: Store permission required }
        "404": { description: Payment not found or mock mode disabled }
        "409": { description: Conflicting identity, amount, destination, or terminal state }
  /stores/{storeId}/mpesa/payments/{orderId}/reconcile:
    post:
      summary: Reconcile uncertain KES collection and disbursement
      security: [{ ApiKey: [CanModifyStoreSettings] }]
      parameters:
        - $ref: "#/components/parameters/StoreId"
        - name: orderId
          in: path
          required: true
          schema: { type: string }
      responses:
        "200": { description: Current state or simulated delivery result; HTTP success alone does not mean payment success }
        "401": { description: Authentication required }
        "403": { description: Store permission required }
        "404": { description: Payment not found or mock mode disabled }
        "409": { description: Conflicting identity, amount, destination, or terminal state }
  /stores/{storeId}/mpesa/payments/{orderId}/mock-callback:
    post:
      summary: Simulate an authenticated Development-only provider observation
      security: [{ ApiKey: [CanModifyStoreSettings] }]
      parameters:
        - $ref: "#/components/parameters/StoreId"
        - name: orderId
          in: path
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [amountKes, destination, collection, disbursement]
              properties:
                amountKes: { type: number }
                destination: { type: string }
                collection: { type: string, enum: [Unknown, Pending, Succeeded, Failed] }
                disbursement: { type: string, enum: [Unknown, Pending, Succeeded, Failed] }
                deliver: { type: boolean, default: true }
      responses:
        "200": { description: Current state or simulated delivery result; HTTP success alone does not mean payment success }
        "401": { description: Authentication required }
        "403": { description: Store permission required }
        "404": { description: Payment not found or mock mode disabled }
        "409": { description: Conflicting identity, amount, destination, or terminal state }
...
components:
  securitySchemes:
    ApiKey:
      type: apiKey
      in: header
      name: Authorization
      description: >
        BTCPay Greenfield API key, sent as `Authorization: token <api-key>`.
        Required permission depends on the endpoint: subscription/signup
        endpoints need an unscoped `btcpay.store.canmodifystoresettings`
        (CanModifyStoreSettingsUnscoped) key; all `/stores/{storeId}/...`
        endpoints need that permission scoped to the specific store.
  parameters:
    StoreId:
      name: storeId
      in: path
      required: true
      schema: { type: string }
...
  schemas:
...
    TandoMpesaDestinationType:
      type: string
      enum: [MobileNumber, TillNumber, PayBill]
    TandoMpesaSettingsRequest:
      type: object
      properties:
        destinationType: { $ref: "#/components/schemas/TandoMpesaDestinationType" }
        destination:
          type: string
          nullable: true
          description: Kenyan MSISDN for MobileNumber, or 5-10 digit number for TillNumber/PayBill.
        accountNumber:
          type: string
          nullable: true
          description: Optional, PayBill only.
...